Throw exception if deserialize json failed
