+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/impl/UsersServiceImpl.java
@@ -28,6 +28,7 @@ import org.apache.dolphinscheduler.api.utils.CheckUtils;
 import org.apache.dolphinscheduler.api.utils.PageInfo;
 import org.apache.dolphinscheduler.api.utils.Result;
 import org.apache.dolphinscheduler.common.constants.Constants;
+import org.apache.dolphinscheduler.common.constants.SystemConstants;
 import org.apache.dolphinscheduler.common.enums.AuthorizationType;
 import org.apache.dolphinscheduler.common.enums.Flag;
 import org.apache.dolphinscheduler.common.enums.UserType;
@@ -63,7 +64,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
-import java.util.TimeZone;
 import java.util.stream.Collectors;
 
 import lombok.extern.slf4j.Slf4j;
@@ -891,7 +891,7 @@ public class UsersServiceImpl extends BaseServiceImpl implements UsersService {
 
         // add system default timezone if not user timezone
         if (StringUtils.isEmpty(user.getTimeZone())) {
-            user.setTimeZone(TimeZone.getDefault().toZoneId().getId());
+            user.setTimeZone(SystemConstants.DEFAULT_TIME_ZONE.toZoneId().getId());
         }
 
         // remove password
